Move publication updates into an UpdateHandler
Having to manually add a function reference everywhere it could
be possibly relevant is super error prone. Therefore, an UpdateHandler
that is triggered everytime a snapshot is taken is hopefully more
suited for this task. Thanks Julian.
